Randomization implemented but not correct -- see below
On x86-64 the compiler will encode displacement values as single bytes
if they fall within -128 -> 127.  This limits the amount of
randomization, as changing the displacement to a value outside this
range increases the size of the encoded instruction and thus affects
code pointer targets.  Chameleon still needs to take this limitation
into account.