Several updates, but not quite stable -- see below
This commit removes all reliance on us injecting system calls or
modifying the kernel to get userfaultfd file descriptors across
execve() boundaries.  Instead, this commit uses compel to execute
system calls in the context of the child and also provides a parasite
for creating and passing a userfaultfd file descriptor inside the child
after execve().  This commit isn't stable:

- Strange implicit declaration warning about fds_send_fd in parasite
- Need 2 calls to get child in compel-controllable state
- Floating point exception in printf_core