Implemented advancing child threads to transformation points
When the timer triggers a re-randomization (or whatever policy
instructs us to re-randomize), we may stop threads at points where we
don't have transformation metadata.  To fix this, we spray the code of
the function where the thread is currently executing with breakpoints
at transformation instructions and advance the thread to one of those
points.

To do: we can potentially add more transformation points, e.g., inside
the headers of loops.  If we do so, we need to detect these additional
points and add them to the list of locations to add breakpoints.