Cleanly handle when spraying breakpoints fails
Occasionally when trying to advance a child to a transformation point,
ptrace() returns EIO.  This seems to happen because we're trying to
read/write a page that isn't mapped, and ptrace() won't force the page to
get mapped in under the hood (maybe because we're handling those faults
via userfaultfd?).  Rather than dying, restore any previously sprayed
breakpoints and skip that re-randomization.